feat: [US-027] - Fix clippy warnings: dead code and simple lints
Fix all simple clippy warnings across gpui-desktop and sandpaper-core:
- Add #[allow(dead_code)] for intentionally unused items (future API surface)
- Remove truly unused code (SettingsModalModel, clear_plugin_error)
- Fix useless_conversion: remove redundant .into_iter() on splice() calls
- Fix collapsed_if: merge nested conditions
- Fix question_mark: use ? operator instead of manual match/if-let-else
- Fix redundant_guards: simplify match arms (Some(ttl) if ttl == 0 → Some(0))
- Fix manual_div_ceil: use .div_ceil() method
- Fix while_let_on_iterator: use for loops
- Fix manual_strip: use .strip_prefix()
- Fix manual_clamp: use .clamp()
- Fix needless_borrow, let_and_return, unnecessary_cast
- Fix type_complexity with ClickHandler type alias
- Fix approx_constant in test (3.14 → 2.72)
- Fix useless_vec, needless_return, iter_cloned_collect
- Remove unused PathBuf import

Only too_many_arguments and result_large_err remain (reserved for US-028).
21 files changed, 106 insertions(+), 168 deletions(-).

+++ apps/gpui-desktop/src/ui/components/skeleton.rs
@@ -22,6 +22,7 @@ impl Skeleton {
         }
     }
 
+    #[allow(dead_code)] // Public API for custom radius
     pub(crate) fn radius(mut self, radius: Pixels) -> Self {
         self.radius = radius;
         self
@@ -45,7 +46,7 @@ impl RenderOnce for Skeleton {
                     .with_easing(ease_in_out),
                 |el, delta| {
                     // Oscillate opacity between 0.4 and 1.0
-                    let opacity = 0.4 + 0.6 * (1.0 - (delta * 2.0 - 1.0).abs()) as f32;
+                    let opacity = 0.4 + 0.6 * (1.0 - (delta * 2.0 - 1.0).abs());
                     el.opacity(opacity)
                 },
             )
